+++ dart/lib/src/sentry_options.dart
@@ -91,6 +91,10 @@ class SentryOptions {
     _maxSpans = maxSpans;
   }
 
+  /// Configures up to which size request bodies should be included in events.
+  /// This does not change whether an event is captured.
+  MaxRequestBodySize maxRequestBodySize = MaxRequestBodySize.never;
+
   SentryLogger _logger = noOpLogger;
 
   /// Logger interface to log useful debugging information if debug is enabled
@@ -227,6 +231,18 @@ class SentryOptions {
 
   /// Whether to send personal identifiable information along with events
   bool sendDefaultPii = false;
+
+  /// Configures whether to record exceptions for failed requests.
+  /// Examples for captures exceptions are:
+  /// - In an browser environment this can be requests which fail because of CORS.
+  /// - In an mobile or desktop application this can be requests which failed
+  ///   because the connection was interrupted.
+  /// Use with [SentryHttpClient] or [Dio] integration for this to work
+  bool captureFailedHttpRequests = false;
+
+  /// Whether to records requests as breadcrumbs. This is on by default.
+  /// It only has an effect when the SentryHttpClient or dio integration is in use
+  bool recordHttpBreadcrumbs = true;
 
   /// Whether [SentryEvent] deduplication is enabled.
   /// Can be further configured with [maxDeduplicationItems].