+++ dart/test/http_client/sentry_http_client_test.dart
@@ -44,10 +44,10 @@ void main() {
 
     test('one captured event with when enabling $FailedRequestClient',
         () async {
+      fixture.hub.options.captureFailedHttpRequests = true;
+      fixture.hub.options.recordHttpBreadcrumbs = true;
       final sut = fixture.getSut(
         client: createThrowingClient(),
-        captureFailedRequests: true,
-        recordBreadcrumbs: true,
       );
 
       await expectLater(() async => await sut.get(requestUri), throwsException);
@@ -74,10 +74,9 @@ void main() {
     });
 
     test('no captured span if tracing disabled', () async {
+      fixture.hub.options.recordHttpBreadcrumbs = false;
       final sut = fixture.getSut(
         client: fixture.getClient(statusCode: 200, reason: 'OK'),
-        recordBreadcrumbs: false,
-        networkTracing: false,
       );
 
       final response = await sut.get(requestUri);
@@ -87,10 +86,10 @@ void main() {
     });
 
     test('captured span if tracing enabled', () async {
+      fixture.hub.options.tracesSampleRate = 1.0;
+      fixture.hub.options.recordHttpBreadcrumbs = false;
       final sut = fixture.getSut(
         client: fixture.getClient(statusCode: 200, reason: 'OK'),
-        recordBreadcrumbs: false,
-        networkTracing: true,
       );
 
       final response = await sut.get(requestUri);
@@ -115,21 +114,13 @@ class CloseableMockClient extends Mock implements BaseClient {}
 class Fixture {
   SentryHttpClient getSut({
     MockClient? client,
-    bool captureFailedRequests = false,
-    MaxRequestBodySize maxRequestBodySize = MaxRequestBodySize.never,
     List<SentryStatusCode> badStatusCodes = const [],
-    bool recordBreadcrumbs = true,
-    bool networkTracing = false,
   }) {
     final mc = client ?? getClient();
     return SentryHttpClient(
       client: mc,
       hub: hub,
-      captureFailedRequests: captureFailedRequests,
       failedRequestStatusCodes: badStatusCodes,
-      maxRequestBodySize: maxRequestBodySize,
-      recordBreadcrumbs: recordBreadcrumbs,
-      networkTracing: networkTracing,
     );
   }