+++ dio/lib/src/sentry_dio_client_adapter.dart
@@ -27,14 +27,13 @@ class SentryDioClientAdapter extends HttpClientAdapter {
   SentryDioClientAdapter({
     required HttpClientAdapter client,
     Hub? hub,
-    bool recordBreadcrumbs = true,
-    bool networkTracing = true,
   }) {
     _hub = hub ?? HubAdapter();
 
     var innerClient = client;
 
-    if (networkTracing) {
+    // ignore: invalid_use_of_internal_member
+    if (_hub.options.isTracingEnabled()) {
       innerClient = TracingClientAdapter(client: innerClient, hub: _hub);
     }
 
@@ -42,7 +41,8 @@ class SentryDioClientAdapter extends HttpClientAdapter {
     // We don't want to include the breadcrumbs for the current request
     // when capturing it as a failed request.
     // However it still should be added for following events.
-    if (recordBreadcrumbs) {
+    // ignore: invalid_use_of_internal_member
+    if (_hub.options.recordHttpBreadcrumbs) {
       innerClient = BreadcrumbClientAdapter(client: innerClient, hub: _hub);
     }